+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Add `Alchemy#statPath(path, options)` function
 * Allow `Model.addField()` to override already existing fields
 * Add pledge support to `Alchemy#openUrl()`
+* `Alchemy#findPathToBinarySync()` will now also look in the `/bin/` directory
 
 ## 1.0.6 (2018-12-06)